requirements.txt
# Core
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
scipy>=1.11
statsmodels>=0.14
matplotlib>=3.8
joblib>=1.3
pyyaml>=6.0.1
pydantic>=2.6
requests>=2.31
duckdb>=1.0.0

# Data sources
pybaseball>=2.2.7
meteostat>=1.6.8

# Explainability
shap>=0.45.1

# Dev & tests
pytest>=8.0
pytest-cov>=4.1
python-dateutil>=2.9
tqdm>=4.66

# Notebooks & parquet
jupyterlab>=4.0
pyarrow>=15.0
ipykernel>=6.29
